# docs/openapi/a_memorix_v1.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "a_memorix/api/v1/auth.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/namespaces/{context.namespaceId}/memories:batchIngest": {
      "post": {
        "operationId": "MemoryService_BatchIngestText",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1BatchIngestTextResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "context.namespaceId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MemoryServiceBatchIngestTextBody"
            }
          }
        ],
        "tags": [
          "MemoryService"
        ]
      }
    },
    "/v1/namespaces/{context.namespaceId}/memories:delete": {
      "post": {
        "operationId": "MemoryService_DeleteMemory",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteMemoryResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "context.namespaceId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MemoryServiceDeleteMemoryBody"
            }
          }
        ],
        "tags": [
          "MemoryService"
        ]
      }
    },
    "/v1/namespaces/{context.namespaceId}/memories:deleteBySource": {
      "post": {
        "operationId": "JobService_SubmitDeleteBySource",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SubmitDeleteBySourceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "context.namespaceId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/JobServiceSubmitDeleteBySourceBody"
            }
          }
        ],
        "tags": [
          "JobService"
        ]
      }
    },
    "/v1/namespaces/{context.namespaceId}/memories:get": {
      "post": {
        "operationId": "MemoryService_GetMemory",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetMemoryResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "context.namespaceId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MemoryServiceGetMemoryBody"
            }
          }
        ],
        "tags": [
          "MemoryService"
        ]
      }
    },
    "/v1/namespaces/{context.namespaceId}/memories:ingest": {
      "post": {
        "operationId": "MemoryService_IngestText",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1IngestTextResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "context.namespaceId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MemoryServiceIngestTextBody"
            }
          }
        ],
        "tags": [
          "MemoryService"
        ]
      }
    },
    "/v1/namespaces/{context.namespaceId}/memories:search": {
      "post": {
        "operationId": "MemoryService_SearchMemory",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SearchMemoryResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "context.namespaceId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MemoryServiceSearchMemoryBody"
            }
          }
        ],
        "tags": [
          "MemoryService"
        ]
      }
    },
...
    "/v1/namespaces/{namespaceId}/jobs": {
      "get": {
        "operationId": "JobService_ListJobs",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListJobsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "namespaceId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "pageToken",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "JobService"
        ]
      }
    },
    "/v1/namespaces/{namespaceId}/jobs/{jobId}": {
      "get": {
        "operationId": "JobService_GetJob",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetJobResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "namespaceId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "JobService"
        ]
      }
    },
    "/v1/namespaces/{namespaceId}/jobs/{jobId}:cancel": {
      "post": {
        "operationId": "JobService_CancelJob",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CancelJobResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "namespaceId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "JobService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "JobServiceSubmitDeleteBySourceBody": {
      "type": "object",
      "properties": {
        "context": {
          "type": "object",
          "properties": {
            "agentId": {
              "type": "string"
            },
            "principalId": {
              "type": "string"
            },
            "conversationId": {
              "type": "string"
            },
            "userId": {
              "type": "string"
            },
            "groupId": {
              "type": "string"
            },
            "requestId": {
              "type": "string"
            },
            "traceId": {
              "type": "string"
            },
            "idempotencyKey": {
              "type": "string"
            }
          },
          "description": "RequestContext carries namespace identity and request correlation data."
        },
        "source": {
          "type": "string"
        },
        "reason": {
          "type": "string"
        }
      }
    },
    "MemoryServiceBatchIngestTextBody": {
      "type": "object",
      "properties": {
        "context": {
          "type": "object",
          "properties": {
            "agentId": {
              "type": "string"
            },
            "principalId": {
              "type": "string"
            },
            "conversationId": {
              "type": "string"
            },
            "userId": {
              "type": "string"
            },
            "groupId": {
              "type": "string"
            },
            "requestId": {
              "type": "string"
            },
            "traceId": {
              "type": "string"
            },
            "idempotencyKey": {
              "type": "string"
            }
          },
          "description": "RequestContext carries namespace identity and request correlation data."
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1IngestTextInput"
          }
        }
      }
    },
    "MemoryServiceDeleteMemoryBody": {
      "type": "object",
      "properties": {
        "context": {
          "type": "object",
          "properties": {
            "agentId": {
              "type": "string"
            },
            "principalId": {
              "type": "string"
            },
            "conversationId": {
              "type": "string"
            },
            "userId": {
              "type": "string"
            },
            "groupId": {
              "type": "string"
            },
            "requestId": {
              "type": "string"
            },
            "traceId": {
              "type": "string"
            },
            "idempotencyKey": {
              "type": "string"
            }
          },
          "description": "RequestContext carries namespace identity and request correlation data."
        },
        "memoryId": {
          "type": "string"
        },
        "externalId": {
          "type": "string"
        },
        "reason": {
          "type": "string"
        }
      }
    },
...
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1BatchIngestTextResponse": {
      "type": "object",
      "properties": {
        "results": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1BatchIngestItemResult"
          }
        },
        "succeeded": {
          "type": "integer",
          "format": "int64"
        },
        "failed": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
    "v1DeleteMemoryResponse": {
      "type": "object",
      "properties": {
        "operationId": {
          "type": "string"
        },
        "deletedCount": {
          "type": "integer",
          "format": "int64"
        },
        "deletedMemoryIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "v1GetMemoryResponse": {
      "type": "object",
      "properties": {
        "memory": {
          "$ref": "#/definitions/v1MemoryRecord"
        }
      }
    },
...
    "v1SubmitDeleteBySourceResponse": {
      "type": "object",
      "properties": {
        "job": {
          "$ref": "#/definitions/v1JobInfo"
        }
      }
    },
...
  }
}